+++ packages/blaxel/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @computesdk/blaxel
 
+## 1.2.8
+
+### Patch Changes
+
+- a146b97: Adding in proper background command via background: true
+- Updated dependencies [a146b97]
+  - computesdk@1.8.8
+
+## 1.2.7
+
+### Patch Changes
+
+- Updated dependencies [04ffecf]
+  - computesdk@1.8.7
+
 ## 1.2.6
 
 ### Patch Changes